backend/requirements.txt
flask
flask-cors
numpy==1.26.4
torch
transformers
pillow
gunicorn
